Refactor utils into a new file
