+++ package.json
@@ -114,7 +114,7 @@
 		"@types/lodash": "4.14.116",
 		"@types/opn": "5.1.0",
 		"eventsource": "^1.0.5",
-		"ldclient-node": "5.1.1",
+		"ldclient-node": "5.4.2",
 		"lodash.kebabcase": "4.1.1",
 		"opn": "5.3.0",
 		"request": "2.88.0",
